+++ lib/widgets/theme_petal_selector/theme_petal_painter.dart
@@ -35,6 +35,12 @@ class ThemePetalPainter extends CustomPainter {
 
     final Paint paint = Paint()..style = PaintingStyle.fill;
 
+    paint.color = Colors.white10;
+
+    canvas.drawRect(-size.center(Offset.zero) & size, paint);
+
+    final double r = size.width * 0.5;
+
     final Paint ringPaint = Paint()
       ..style = PaintingStyle.stroke
       ..strokeCap = StrokeCap.round
@@ -43,8 +49,8 @@ class ThemePetalPainter extends CustomPainter {
       ..color = selectedColor;
 
     final double petalCenterDistance =
-        size.width * (petalRadiusProc + centerSpaceRadiusProc);
-    final double petalRadius = size.width * petalRadiusProc;
+        r * (petalRadiusProc + centerSpaceRadiusProc);
+    final double petalRadius = r * petalRadiusProc;
     final double ringRadius = petalRadius + 4.0;
 
     final double angleDelta = math.pi * 2 / colors.length;